+++ src/App.tsx
@@ -32,7 +32,6 @@ const router = createBrowserRouter([
       { index: true, element: <MainPage /> },
       { path: '/sign-up', element: <SignupPage /> },
       { path: '/password', element: <PasswordPage /> },
-      { path: '/apply', element: <ApplyPage /> },
       { path: '/complete', element: <CompletePage /> },
       { path: '/my', element: <MyPage /> },
       { path: '/result', element: <ResultPage /> },
